+++ src/commands/plugin.ts
@@ -30,7 +30,7 @@ export async function switchEndpoint(): Promise<void> {
     if (!choice || choice.value === getLeetCodeEndpoint()) {
         return;
     }
-    const leetCodeConfig: vscode.WorkspaceConfiguration = vscode.workspace.getConfiguration("leetcode");
+    const leetCodeConfig: vscode.WorkspaceConfiguration = vscode.workspace.getConfiguration("lovecode");
     try {
         const endpoint: string = choice.value;
         await leetCodeExecutor.switchEndpoint(endpoint);
@@ -41,7 +41,7 @@ export async function switchEndpoint(): Promise<void> {
     }
 
     try {
-        await vscode.commands.executeCommand("leetcode.signout");
+        await vscode.commands.executeCommand("lovecode.signout");
         await deleteCache();
         await promptForSignIn();
     } catch (error) {
@@ -50,7 +50,7 @@ export async function switchEndpoint(): Promise<void> {
 }
 
 export function getLeetCodeEndpoint(): string {
-    const leetCodeConfig: vscode.WorkspaceConfiguration = vscode.workspace.getConfiguration("leetcode");
+    const leetCodeConfig: vscode.WorkspaceConfiguration = vscode.workspace.getConfiguration("lovecode");
     return leetCodeConfig.get<string>("endpoint", Endpoint.LeetCode);
 }
 
@@ -77,12 +77,12 @@ export async function switchSortingStrategy(): Promise<void> {
         return;
     }
 
-    const leetCodeConfig: vscode.WorkspaceConfiguration = vscode.workspace.getConfiguration("leetcode");
+    const leetCodeConfig: vscode.WorkspaceConfiguration = vscode.workspace.getConfiguration("lovecode");
     await leetCodeConfig.update("problems.sortStrategy", choice.value, true);
     await leetCodeTreeDataProvider.refresh();
 }
 
 export function getSortingStrategy(): SortingStrategy {
-    const leetCodeConfig: vscode.WorkspaceConfiguration = vscode.workspace.getConfiguration("leetcode");
+    const leetCodeConfig: vscode.WorkspaceConfiguration = vscode.workspace.getConfiguration("lovecode");
     return leetCodeConfig.get<SortingStrategy>("problems.sortStrategy", SortingStrategy.None);
 }