+++ src/explorer/LeetCodeTreeDataProvider.ts
@@ -32,7 +32,7 @@ export class LeetCodeTreeDataProvider implements vscode.TreeDataProvider<LeetCod
                 label: element.name,
                 collapsibleState: vscode.TreeItemCollapsibleState.None,
                 command: {
-                    command: "leetcode.signin",
+                    command: "lovecode.signin",
                     title: "Sign in to Lode",
                 },
             };