+++ src/statusbar/LeetCodeStatusBarItem.ts
@@ -9,7 +9,7 @@ export class LeetCodeStatusBarItem implements vscode.Disposable {
 
     constructor() {
         this.statusBarItem = vscode.window.createStatusBarItem();
-        this.statusBarItem.command = "leetcode.manageSessions";
+        this.statusBarItem.command = "lovecode.manageSessions";
     }
 
     public updateStatusBar(status: UserStatus, user?: string): void {