+++ src/utils/uiUtils.ts
@@ -44,7 +44,7 @@ export async function promptForSignIn(): Promise<void> {
     );
     switch (choice) {
         case DialogOptions.yes:
-            await vscode.commands.executeCommand("leetcode.signin");
+            await vscode.commands.executeCommand("lovecode.signin");
             break;
         case DialogOptions.singUp:
             if (getLeetCodeEndpoint()) {