+++ src/webview/leetCodePreviewProvider.ts
@@ -9,7 +9,7 @@ import { markdownEngine } from "./markdownEngine";
 
 class LeetCodePreviewProvider extends LeetCodeWebview {
 
-    protected readonly viewType: string = "leetcode.preview";
+    protected readonly viewType: string = "lovecode.preview";
     private node: IProblem;
     private description: IDescription;
     private sideMode: boolean = false;
@@ -136,7 +136,7 @@ class LeetCodePreviewProvider extends LeetCodeWebview {
     protected async onDidReceiveMessage(message: IWebViewMessage): Promise<void> {
         switch (message.command) {
             case "ShowProblem": {
-                await commands.executeCommand("leetcode.showProblem", this.node);
+                await commands.executeCommand("lovecode.showProblem", this.node);
                 break;
             }
         }